+++ src/Database/Reader.php
@@ -34,6 +34,7 @@ use MaxMind\Db\Reader\InvalidDatabaseException;
 class Reader implements ProviderInterface
 {
     private $dbReader;
+    private $dbType;
     private $locales;
 
     /**
@@ -51,6 +52,7 @@ class Reader implements ProviderInterface
         $locales = ['en']
     ) {
         $this->dbReader = new DbReader($filename);
+        $this->dbType = $this->dbReader->metadata()->databaseType;
         $this->locales = $locales;
     }
 
@@ -212,9 +214,11 @@ class Reader implements ProviderInterface
 
     private function modelFor($class, $type, $ipAddress)
     {
-        $record = $this->getRecord($class, $type, $ipAddress);
+        list($record, $prefixLen) = $this->getRecord($class, $type, $ipAddress);
 
         $record['traits']['ip_address'] = $ipAddress;
+        $record['traits']['prefix_len'] = $prefixLen;
+
         $class = 'GeoIp2\\Model\\' . $class;
 
         return new $class($record, $this->locales);
@@ -222,9 +226,10 @@ class Reader implements ProviderInterface
 
     private function flatModelFor($class, $type, $ipAddress)
     {
-        $record = $this->getRecord($class, $type, $ipAddress);
+        list($record, $prefixLen) = $this->getRecord($class, $type, $ipAddress);
 
         $record['ip_address'] = $ipAddress;
+        $record['prefix_len'] = $prefixLen;
         $class = 'GeoIp2\\Model\\' . $class;
 
         return new $class($record);
@@ -232,14 +237,13 @@ class Reader implements ProviderInterface
 
     private function getRecord($class, $type, $ipAddress)
     {
-        if (strpos($this->metadata()->databaseType, $type) === false) {
+        if (strpos($this->dbType, $type) === false) {
             $method = lcfirst($class);
             throw new \BadMethodCallException(
-                "The $method method cannot be used to open a "
-                . $this->metadata()->databaseType . ' database'
+                "The $method method cannot be used to open a {$this->dbType} database"
             );
         }
-        $record = $this->dbReader->get($ipAddress);
+        list($record, $prefixLen) = $this->dbReader->getWithPrefixLen($ipAddress);
         if ($record === null) {
             throw new AddressNotFoundException(
                 "The address $ipAddress is not in the database."
@@ -259,7 +263,7 @@ class Reader implements ProviderInterface
             );
         }
 
-        return $record;
+        return [$record, $prefixLen];
     }
 
     /**